+++ js/loader/ScriptLoadRequest.h
@@ -37,6 +37,7 @@ namespace mozilla::dom {
 class ScriptLoadContext;
 class WorkerLoadContext;
 class WorkletLoadContext;
+enum class RequestPriority : uint8_t;
 
 }  // namespace mozilla::dom
 
@@ -88,6 +89,7 @@ class ScriptFetchOptions {
   ScriptFetchOptions(mozilla::CORSMode aCORSMode,
                      enum mozilla::dom::ReferrerPolicy aReferrerPolicy,
                      const nsAString& aNonce,
+                     mozilla::dom::RequestPriority aFetchPriority,
                      const ParserMetadata aParserMetadata,
                      nsIPrincipal* aTriggeringPrincipal,
                      mozilla::dom::Element* aElement = nullptr);
@@ -110,6 +112,11 @@ class ScriptFetchOptions {
    * fetching any imported modules.
    */
   const nsString mNonce;
+
+  /*
+   * <https://html.spec.whatwg.org/multipage/webappapis.html#script-fetch-options>.
+   */
+  const mozilla::dom::RequestPriority mFetchPriority;
 
   /*
    * The parser metadata used for the initial fetch and for fetching any
@@ -242,16 +249,10 @@ class ScriptLoadRequest
     mScriptData.reset();
   }
 
-  bool IsUTF8ParsingEnabled();
-
   void SetTextSource() {
     MOZ_ASSERT(IsUnknownDataType());
     mDataType = DataType::eTextSource;
-    if (IsUTF8ParsingEnabled()) {
-      mScriptData.emplace(VariantType<ScriptTextBuffer<Utf8Unit>>());
-    } else {
-      mScriptData.emplace(VariantType<ScriptTextBuffer<char16_t>>());
-    }
+    mScriptData.emplace(VariantType<ScriptTextBuffer<Utf8Unit>>());
   }
 
   // Use a vector backed by the JS allocator for script text so that contents
@@ -292,6 +293,10 @@ class ScriptLoadRequest
     MOZ_ASSERT(IsTextSource());
     return IsUTF16Text() ? ScriptText<char16_t>().clearAndFree()
                          : ScriptText<Utf8Unit>().clearAndFree();
+  }
+
+  mozilla::dom::RequestPriority FetchPriority() const {
+    return mFetchOptions->mFetchPriority;
   }
 
   enum mozilla::dom::ReferrerPolicy ReferrerPolicy() const {