+++ layout/style/nsStyleStruct.h
@@ -617,6 +617,8 @@ struct MOZ_NEEDS_MEMMOVABLE_MEMBERS nsStyleOutline {
     return false;
   }
 
+  nsSize EffectiveOffsetFor(const nsRect& aRect) const;
+
  protected:
   // The actual value of outline-width is the computed value (an absolute
   // length, forced to zero when outline-style is none) rounded to device
@@ -1196,7 +1198,7 @@ struct StyleScrollTimeline {
   // elements.
   void SetInitialValues() {}
 
-  const nsAtom* GetName() const { return mName._0.AsAtom(); }
+  nsAtom* GetName() const { return mName._0.AsAtom(); }
   StyleScrollAxis GetAxis() const { return mAxis; }
 
   bool operator==(const StyleScrollTimeline& aOther) const {
@@ -1220,7 +1222,7 @@ struct StyleViewTimeline {
   // elements.
   void SetInitialValues() {}
 
-  const nsAtom* GetName() const { return mName._0.AsAtom(); }
+  nsAtom* GetName() const { return mName._0.AsAtom(); }
   StyleScrollAxis GetAxis() const { return mAxis; }
   const StyleViewTimelineInset& GetInset() const { return mInset; }
 
@@ -1300,7 +1302,7 @@ struct MOZ_NEEDS_MEMMOVABLE_MEMBERS nsStyleDisplay {
 
   mozilla::StyleBackfaceVisibility mBackfaceVisibility;
   mozilla::StyleTransformStyle mTransformStyle;
-  mozilla::StyleGeometryBox mTransformBox;
+  mozilla::StyleTransformBox mTransformBox;
 
   mozilla::StyleTransform mTransform;
   mozilla::StyleRotate mRotate;