+++ toolkit/components/translations/tests/browser/shared-head.js
@@ -455,6 +455,7 @@ async function loadTestPage({
       ["browser.translations.enable", true],
       ["browser.translations.logLevel", "All"],
       ["browser.translations.panelShown", true],
+      ["browser.translations.automaticallyPopup", true],
       ...(prefs ?? []),
     ],
   });
@@ -487,7 +488,7 @@ async function loadTestPage({
   BrowserTestUtils.loadURIString(tab.linkedBrowser, page);
   await BrowserTestUtils.browserLoaded(tab.linkedBrowser);
 
-  if (autoOffer) {
+  if (autoOffer && TranslationsParent.shouldAlwaysOfferTranslations()) {
     info("Waiting for the popup to be automatically shown.");
     await waitForCondition(() => {
       const panel = document.getElementById("translations-panel");